requirements.txt
python-telegram-bot==20.0
python-docx==0.8.11
openpyxl==3.0.10
